Move SwiftASTContext & TypeSystemSwift* into Plugins/TypeSystem/Swift
to mirror a similar move spearheaded by the Clang typesystem.
