+++ lldb/source/Core/ValueObject.cpp
@@ -54,7 +54,7 @@
 #include "Plugins/ExpressionParser/Clang/ClangModulesDeclVendor.h"
 
 // BEGIN SWIFT
-#include "lldb/Symbol/SwiftASTContext.h"
+#include "Plugins/TypeSystem/Swift/SwiftASTContext.h"
 // END SWIFT
 
 #include "llvm/Support/Compiler.h"
