+++ lldb/source/Plugins/ExpressionParser/Swift/CMakeLists.txt
@@ -26,6 +26,7 @@ add_lldb_library(lldbPluginExpressionParserSwift PLUGIN
     lldbSymbol
     lldbTarget
     lldbUtility
+    lldbPluginTypeSystemSwift
     swiftAST
     swiftBasic
     swiftClangImporter
