+++ lldb/source/Plugins/Language/Swift/SwiftDictionary.cpp
@@ -12,9 +12,9 @@
 
 #include "SwiftDictionary.h"
 
-#include "lldb/DataFormatters/FormattersHelpers.h"
 #include "Plugins/TypeSystem/Clang/TypeSystemClang.h"
-#include "lldb/Symbol/SwiftASTContext.h"
+#include "Plugins/TypeSystem/Swift/SwiftASTContext.h"
+#include "lldb/DataFormatters/FormattersHelpers.h"
 #include "lldb/Target/Process.h"
 #include "lldb/Target/SwiftLanguageRuntime.h"
