+++ lldb/source/Plugins/TypeSystem/CMakeLists.txt
@@ -1,1 +1,2 @@
 add_subdirectory(Clang)
+add_subdirectory(Swift)
