+++ lldb/source/Plugins/TypeSystem/Swift/SwiftASTContext.h
@@ -2,7 +2,7 @@
 //
 // This source file is part of the Swift.org open source project
 //
-// Copyright (c) 2014 - 2016 Apple Inc. and the Swift project authors
+// Copyright (c) 2014 - 2020 Apple Inc. and the Swift project authors
 // Licensed under Apache License v2.0 with Runtime Library Exception
 //
 // See https://swift.org/LICENSE.txt for license information
@@ -14,29 +14,15 @@
 #define liblldb_SwiftASTContext_h_
 
 #include "Plugins/ExpressionParser/Swift/SwiftPersistentExpressionState.h"
+#include "Plugins/TypeSystem/Swift/TypeSystemSwift.h"
+#include "Plugins/TypeSystem/Swift/TypeSystemSwiftTypeRef.h"
 #include "lldb/Core/ClangForward.h"
 #include "lldb/Core/SwiftForward.h"
 #include "lldb/Core/ThreadSafeDenseMap.h"
 #include "lldb/Core/ThreadSafeDenseSet.h"
-#include "lldb/Symbol/CompilerType.h"
-#include "lldb/Symbol/SymbolFile.h"
-#include "lldb/Symbol/TypeSystem.h"
-#include "lldb/Utility/ConstString.h"
-#include "lldb/lldb-private.h"
-
 #include "lldb/Utility/Either.h"
-#include "lldb/Utility/Status.h"
-
-#include "llvm/ADT/Optional.h"
-#include "llvm/Support/Threading.h"
+#include "llvm/ADT/StringRef.h"
 #include "llvm/Target/TargetOptions.h"
-
-// FIXME: needed only for the DenseMap.
-#include "clang/Basic/Module.h"
-#include "clang/APINotes/APINotesManager.h"
-
-#include <map>
-#include <set>
 
 namespace swift {
 enum class IRGenDebugInfoLevel : unsigned;
@@ -88,437 +74,6 @@ class SwiftASTContext;
 class ClangExternalASTSourceCallbacks;
 
 CompilerType ToCompilerType(swift::Type qual_type);
-
-/// The implementation of lldb::Type's m_payload field for TypeSystemSwift.
-class TypePayloadSwift {
-  /// Layout: bit 1 ... IsFixedValueBuffer.
-  Type::Payload m_payload = 0;
-
-  static constexpr unsigned FixedValueBufferBit = 1;
-
-public:
-  TypePayloadSwift() = default;
-  explicit TypePayloadSwift(bool is_fixed_value_buffer);
-  explicit TypePayloadSwift(Type::Payload opaque_payload)
-      : m_payload(opaque_payload) {}
-  operator Type::Payload() { return m_payload; }
-
-  /// \return whether this is a Swift fixed-size buffer. Resilient variables in
-  /// fixed-size buffers may be indirect depending on the runtime size of the
-  /// type. This is more a property of the value than of its type.
-  bool IsFixedValueBuffer() {
-    return Flags(m_payload).Test(FixedValueBufferBit);
-  }
-  void SetIsFixedValueBuffer(bool is_fixed_value_buffer) {
-    m_payload = is_fixed_value_buffer
-                    ? Flags(m_payload).Set(FixedValueBufferBit)
-                    : Flags(m_payload).Clear(FixedValueBufferBit);
-  }
-};
-
-/// Abstract base class for all Swift TypeSystems.
-///
-/// Swift CompilerTypes are either a mangled name or a Swift AST
-/// type. If the typesystem is a TypeSystemSwiftTypeRef, they are
-/// mangled names.
-///
-/// \verbatim
-///                      TypeSystem (abstract)
-///                            │
-///                            ↓
-///                      TypeSystemSwift (abstract)
-///                        │         │
-///                        ↓         ↓
-///    TypeSystemSwiftTypeRef ⟷ SwiftASTContext (deprecated)
-///                                  │
-///                                  ↓
-///                               SwiftASTContextForExpressions
-///
-/// \endverbatim
-class TypeSystemSwift : public TypeSystem {
-  /// LLVM RTTI support.
-  static char ID;
-
-public:
-  /// LLVM RTTI support.
-  /// \{
-  bool isA(const void *ClassID) const override { return ClassID == &ID; }
-  static bool classof(const TypeSystem *ts) { return ts->isA(&ID); }
-  /// \}
-
-  TypeSystemSwift();
-
-  virtual lldb::TypeSP GetCachedType(ConstString mangled) = 0;
-  virtual void SetCachedType(ConstString mangled,
-                             const lldb::TypeSP &type_sp) = 0;
-  virtual bool IsImportedType(lldb::opaque_compiler_type_t type,
-                              CompilerType *original_type) = 0;
-  virtual CompilerType GetErrorType() = 0;
-  virtual CompilerType GetReferentType(lldb::opaque_compiler_type_t type) = 0;
-  static CompilerType GetInstanceType(CompilerType ct);
-  virtual CompilerType GetInstanceType(lldb::opaque_compiler_type_t type) = 0;
-  enum class TypeAllocationStrategy { eInline, ePointer, eDynamic, eUnknown };
-  virtual TypeAllocationStrategy
-  GetAllocationStrategy(lldb::opaque_compiler_type_t type) = 0;
-  struct TupleElement {
-    ConstString element_name;
-    CompilerType element_type;
-  };
-  virtual CompilerType
-  CreateTupleType(const std::vector<TupleElement> &elements) = 0;
-  virtual void DumpTypeDescription(
-      lldb::opaque_compiler_type_t type, bool print_help_if_available,
-      bool print_extensions_if_available,
-      lldb::DescriptionLevel level = lldb::eDescriptionLevelFull) = 0;
-  virtual void DumpTypeDescription(
-      lldb::opaque_compiler_type_t type, Stream *s,
-      bool print_help_if_available, bool print_extensions_if_available,
-      lldb::DescriptionLevel level = lldb::eDescriptionLevelFull) = 0;
-  virtual CompilerType
-  GetTypeFromMangledTypename(ConstString mangled_typename) = 0;
-
-  /// Unavailable hardcoded functions that don't make sense for Swift.
-  /// \{
-  ConstString DeclContextGetName(void *opaque_decl_ctx) override { return {}; }
-  ConstString DeclContextGetScopeQualifiedName(void *opaque_decl_ctx) override {
-    return {};
-  }
-  bool
-  DeclContextIsClassMethod(void *opaque_decl_ctx,
-                           lldb::LanguageType *language_ptr,
-                           bool *is_instance_method_ptr,
-                           ConstString *language_object_name_ptr) override {
-    return false;
-  }
-  bool IsRuntimeGeneratedType(lldb::opaque_compiler_type_t type) override {
-    return false;
-  }
-  bool IsCharType(lldb::opaque_compiler_type_t type) override { return false; }
-  bool IsCompleteType(lldb::opaque_compiler_type_t type) override {
-    return true;
-  }
-  bool IsConst(lldb::opaque_compiler_type_t type) override { return false; }
-  bool IsCStringType(lldb::opaque_compiler_type_t type,
-                     uint32_t &length) override {
-    return false;
-  }
-  bool IsVectorType(lldb::opaque_compiler_type_t type,
-                    CompilerType *element_type, uint64_t *size) override {
-    return false;
-  }
-  uint32_t IsHomogeneousAggregate(lldb::opaque_compiler_type_t type,
-                                  CompilerType *base_type_ptr) override {
-    return 0;
-  }
-  bool IsBlockPointerType(lldb::opaque_compiler_type_t type,
-                          CompilerType *function_pointer_type_ptr) override {
-    return false;
-  }
-  bool IsPolymorphicClass(lldb::opaque_compiler_type_t type) override {
-    return false;
-  }
-  bool IsBeingDefined(lldb::opaque_compiler_type_t type) override {
-    return false;
-  }
-  bool CanPassInRegisters(const CompilerType &type) override {
-    // FIXME: Implement this. There was an abort() here to figure out which
-    // tests where hitting this code. At least TestSwiftReturns and
-    // TestSwiftStepping were failing because of this Darwin.
-    return false;
-  }
-  unsigned GetTypeQualifiers(lldb::opaque_compiler_type_t type) override {
-    return 0;
-  }
-  CompilerType
-  GetTypeForDecl(lldb::opaque_compiler_type_t opaque_decl) override {
-    llvm_unreachable("GetTypeForDecl not implemented");
-  }
-  CompilerType GetBasicTypeFromAST(lldb::BasicType basic_type) override {
-    return {};
-  }
-  const llvm::fltSemantics &GetFloatTypeSemantics(size_t byte_size) override {
-    // See: https://reviews.llvm.org/D67239. At this time of writing this API
-    // is only used by DumpDataExtractor for the C type system.
-    llvm_unreachable("GetFloatTypeSemantics not implemented.");
-  }
-  lldb::BasicType
-  GetBasicTypeEnumeration(lldb::opaque_compiler_type_t type) override {
-    return lldb::eBasicTypeInvalid;
-  }
-  uint32_t
-  GetNumVirtualBaseClasses(lldb::opaque_compiler_type_t opaque_type) override {
-    return 0;
-  }
-  CompilerType
-  GetVirtualBaseClassAtIndex(lldb::opaque_compiler_type_t opaque_type,
-                             size_t idx, uint32_t *bit_offset_ptr) override {
-    return {};
-  }
-  /// \}
-protected:
-  /// Used in the logs.
-  std::string m_description;
-};
-
-/// A Swift TypeSystem that does not own a swift::ASTContext.
-class TypeSystemSwiftTypeRef : public TypeSystemSwift {
-  /// LLVM RTTI support.
-  static char ID;
-
-public:
-  /// LLVM RTTI support.
-  /// \{
-  bool isA(const void *ClassID) const override {
-    return ClassID == &ID || TypeSystemSwift::isA(ClassID);
-  }
-  static bool classof(const TypeSystem *ts) { return ts->isA(&ID); }
-  /// \}
-
-  TypeSystemSwiftTypeRef(SwiftASTContext *swift_ast_context);
-
-  Module *GetModule() const;
-  swift::CanType GetCanonicalSwiftType(CompilerType compiler_type);
-  swift::Type GetSwiftType(CompilerType compiler_type);
-  CompilerType ReconstructType(CompilerType type);
-  CompilerType
-  GetTypeFromMangledTypename(ConstString mangled_typename) override;
-
-  // PluginInterface functions
-  ConstString GetPluginName() override;
-  uint32_t GetPluginVersion() override;
-
-  bool SupportsLanguage(lldb::LanguageType language) override;
-  Status IsCompatible() override;
-
-  void DiagnoseWarnings(Process &process, Module &module) const override;
-  DWARFASTParser *GetDWARFParser() override;
-  // CompilerDecl functions
-  ConstString DeclGetName(void *opaque_decl) override {
-    return ConstString("");
-  }
-
-  // CompilerDeclContext functions
-  std::vector<CompilerDecl>
-  DeclContextFindDeclByName(void *opaque_decl_ctx, ConstString name,
-                            const bool ignore_imported_decls) override {
-    return {};
-  }
-
-  bool DeclContextIsContainedInLookup(void *opaque_decl_ctx,
-                                      void *other_opaque_decl_ctx) override {
-    if (opaque_decl_ctx == other_opaque_decl_ctx)
-      return true;
-    return false;
-  }
-
-  // Tests
-#ifndef NDEBUG
-  bool Verify(lldb::opaque_compiler_type_t type) override;
-#endif
-  bool IsArrayType(lldb::opaque_compiler_type_t type,
-                   CompilerType *element_type, uint64_t *size,
-                   bool *is_incomplete) override;
-  bool IsAggregateType(lldb::opaque_compiler_type_t type) override;
-  bool IsDefined(lldb::opaque_compiler_type_t type) override;
-  bool IsFloatingPointType(lldb::opaque_compiler_type_t type, uint32_t &count,
-                           bool &is_complex) override;
-  bool IsFunctionType(lldb::opaque_compiler_type_t type,
-                      bool *is_variadic_ptr) override;
-  size_t
-  GetNumberOfFunctionArguments(lldb::opaque_compiler_type_t type) override;
-  CompilerType GetFunctionArgumentAtIndex(lldb::opaque_compiler_type_t type,
-                                          const size_t index) override;
-  bool IsFunctionPointerType(lldb::opaque_compiler_type_t type) override;
-  bool IsIntegerType(lldb::opaque_compiler_type_t type,
-                     bool &is_signed) override;
-  bool IsPossibleDynamicType(lldb::opaque_compiler_type_t type,
-                             CompilerType *target_type, // Can pass NULL
-                             bool check_cplusplus, bool check_objc) override;
-  bool IsPointerType(lldb::opaque_compiler_type_t type,
-                     CompilerType *pointee_type) override;
-  bool IsScalarType(lldb::opaque_compiler_type_t type) override;
-  bool IsVoidType(lldb::opaque_compiler_type_t type) override;
-  // Type Completion
-  bool GetCompleteType(lldb::opaque_compiler_type_t type) override;
-  // AST related queries
-  uint32_t GetPointerByteSize() override;
-  // Accessors
-  ConstString GetTypeName(lldb::opaque_compiler_type_t type) override;
-  ConstString GetDisplayTypeName(lldb::opaque_compiler_type_t type,
-                                 const SymbolContext *sc) override;
-  ConstString GetMangledTypeName(lldb::opaque_compiler_type_t type) override;
-  uint32_t GetTypeInfo(lldb::opaque_compiler_type_t type,
-                       CompilerType *pointee_or_element_clang_type) override;
-  lldb::LanguageType
-  GetMinimumLanguage(lldb::opaque_compiler_type_t type) override;
-  lldb::TypeClass GetTypeClass(lldb::opaque_compiler_type_t type) override;
-
-  // Creating related types
-  CompilerType GetArrayElementType(lldb::opaque_compiler_type_t type,
-                                   uint64_t *stride) override;
-  CompilerType GetCanonicalType(lldb::opaque_compiler_type_t type) override;
-  int GetFunctionArgumentCount(lldb::opaque_compiler_type_t type) override;
-  CompilerType GetFunctionArgumentTypeAtIndex(lldb::opaque_compiler_type_t type,
-                                              size_t idx) override;
-  CompilerType
-  GetFunctionReturnType(lldb::opaque_compiler_type_t type) override;
-  size_t GetNumMemberFunctions(lldb::opaque_compiler_type_t type) override;
-  TypeMemberFunctionImpl
-  GetMemberFunctionAtIndex(lldb::opaque_compiler_type_t type,
-                           size_t idx) override;
-  CompilerType GetPointeeType(lldb::opaque_compiler_type_t type) override;
-  CompilerType GetPointerType(lldb::opaque_compiler_type_t type) override;
-
-  // Exploring the type
-  llvm::Optional<uint64_t>
-  GetBitSize(lldb::opaque_compiler_type_t type,
-             ExecutionContextScope *exe_scope) override;
-  llvm::Optional<uint64_t>
-  GetByteStride(lldb::opaque_compiler_type_t type,
-                ExecutionContextScope *exe_scope) override;
-  lldb::Encoding GetEncoding(lldb::opaque_compiler_type_t type,
-                             uint64_t &count) override;
-  lldb::Format GetFormat(lldb::opaque_compiler_type_t type) override;
-  uint32_t GetNumChildren(lldb::opaque_compiler_type_t type,
-                          bool omit_empty_base_classes,
-                          const ExecutionContext *exe_ctx) override;
-  uint32_t GetNumFields(lldb::opaque_compiler_type_t type) override;
-  CompilerType GetFieldAtIndex(lldb::opaque_compiler_type_t type, size_t idx,
-                               std::string &name, uint64_t *bit_offset_ptr,
-                               uint32_t *bitfield_bit_size_ptr,
-                               bool *is_bitfield_ptr) override;
-  CompilerType GetChildCompilerTypeAtIndex(
-      lldb::opaque_compiler_type_t type, ExecutionContext *exe_ctx, size_t idx,
-      bool transparent_pointers, bool omit_empty_base_classes,
-      bool ignore_array_bounds, std::string &child_name,
-      uint32_t &child_byte_size, int32_t &child_byte_offset,
-      uint32_t &child_bitfield_bit_size, uint32_t &child_bitfield_bit_offset,
-      bool &child_is_base_class, bool &child_is_deref_of_parent,
-      ValueObject *valobj, uint64_t &language_flags) override;
-  uint32_t GetIndexOfChildWithName(lldb::opaque_compiler_type_t type,
-                                   const char *name,
-                                   bool omit_empty_base_classes) override;
-  size_t
-  GetIndexOfChildMemberWithName(lldb::opaque_compiler_type_t type,
-                                const char *name, bool omit_empty_base_classes,
-                                std::vector<uint32_t> &child_indexes) override;
-  size_t GetNumTemplateArguments(lldb::opaque_compiler_type_t type) override;
-  CompilerType GetTypeForFormatters(lldb::opaque_compiler_type_t type) override;
-  LazyBool ShouldPrintAsOneLiner(lldb::opaque_compiler_type_t type,
-                                 ValueObject *valobj) override;
-  bool IsMeaninglessWithoutDynamicResolution(
-      lldb::opaque_compiler_type_t type) override;
-
-  // Dumping types
-#ifndef NDEBUG
-  /// Convenience LLVM-style dump method for use in the debugger only.
-  LLVM_DUMP_METHOD virtual void
-  dump(lldb::opaque_compiler_type_t type) const override;
-#endif
-
-  void DumpValue(lldb::opaque_compiler_type_t type, ExecutionContext *exe_ctx,
-                 Stream *s, lldb::Format format, const DataExtractor &data,
-                 lldb::offset_t data_offset, size_t data_byte_size,
-                 uint32_t bitfield_bit_size, uint32_t bitfield_bit_offset,
-                 bool show_types, bool show_summary, bool verbose,
-                 uint32_t depth) override;
-
-  bool DumpTypeValue(lldb::opaque_compiler_type_t type, Stream *s,
-                     lldb::Format format, const DataExtractor &data,
-                     lldb::offset_t data_offset, size_t data_byte_size,
-                     uint32_t bitfield_bit_size, uint32_t bitfield_bit_offset,
-                     ExecutionContextScope *exe_scope,
-                     bool is_base_class) override;
-
-  void DumpTypeDescription(
-      lldb::opaque_compiler_type_t type,
-      lldb::DescriptionLevel level = lldb::eDescriptionLevelFull) override;
-  void DumpTypeDescription(
-      lldb::opaque_compiler_type_t type, Stream *s,
-      lldb::DescriptionLevel level = lldb::eDescriptionLevelFull) override;
-  void DumpSummary(lldb::opaque_compiler_type_t type, ExecutionContext *exe_ctx,
-                   Stream *s, const DataExtractor &data,
-                   lldb::offset_t data_offset, size_t data_byte_size) override;
-  bool IsPointerOrReferenceType(lldb::opaque_compiler_type_t type,
-                                CompilerType *pointee_type) override;
-  llvm::Optional<size_t>
-  GetTypeBitAlign(lldb::opaque_compiler_type_t type,
-                  ExecutionContextScope *exe_scope) override;
-  CompilerType GetBuiltinTypeForEncodingAndBitSize(lldb::Encoding encoding,
-                                                   size_t bit_size) override {
-    return CompilerType();
-  }
-  bool IsTypedefType(lldb::opaque_compiler_type_t type) override;
-  CompilerType GetTypedefedType(lldb::opaque_compiler_type_t type) override;
-  CompilerType
-  GetFullyUnqualifiedType(lldb::opaque_compiler_type_t type) override;
-  CompilerType GetNonReferenceType(lldb::opaque_compiler_type_t type) override;
-  CompilerType
-  GetLValueReferenceType(lldb::opaque_compiler_type_t type) override;
-  CompilerType
-  GetRValueReferenceType(lldb::opaque_compiler_type_t type) override;
-  uint32_t GetNumDirectBaseClasses(lldb::opaque_compiler_type_t type) override;
-  CompilerType GetDirectBaseClassAtIndex(lldb::opaque_compiler_type_t type,
-                                         size_t idx,
-                                         uint32_t *bit_offset_ptr) override;
-  bool IsReferenceType(lldb::opaque_compiler_type_t type,
-                       CompilerType *pointee_type, bool *is_rvalue) override;
-  bool
-  ShouldTreatScalarValueAsAddress(lldb::opaque_compiler_type_t type) override;
-
-  // Swift-specific methods.
-  lldb::TypeSP GetCachedType(ConstString mangled) override;
-  void SetCachedType(ConstString mangled, const lldb::TypeSP &type_sp) override;
-  bool IsImportedType(lldb::opaque_compiler_type_t type,
-                      CompilerType *original_type) override;
-  CompilerType GetErrorType() override;
-  CompilerType GetReferentType(lldb::opaque_compiler_type_t type) override;
-  CompilerType GetInstanceType(lldb::opaque_compiler_type_t type) override;
-  TypeAllocationStrategy
-  GetAllocationStrategy(lldb::opaque_compiler_type_t type) override;
-  CompilerType
-  CreateTupleType(const std::vector<TupleElement> &elements) override;
-  void DumpTypeDescription(
-      lldb::opaque_compiler_type_t type, bool print_help_if_available,
-      bool print_extensions_if_available,
-      lldb::DescriptionLevel level = lldb::eDescriptionLevelFull) override;
-  void DumpTypeDescription(
-      lldb::opaque_compiler_type_t type, Stream *s,
-      bool print_help_if_available, bool print_extensions_if_available,
-      lldb::DescriptionLevel level = lldb::eDescriptionLevelFull) override;
-
-private:
-  /// Helper that creates an AST type from \p type.
-  void *ReconstructType(lldb::opaque_compiler_type_t type);
-  /// Cast \p opaque_type as a mangled name.
-  const char *AsMangledName(lldb::opaque_compiler_type_t opaque_type);
-
-  /// Wrap \p node as \p Global(TypeMangling(node)), remangle the type
-  /// and create a CompilerType from it.
-  CompilerType RemangleAsType(swift::Demangle::Demangler &Dem,
-                              swift::Demangle::NodePointer node);
-
-  /// Demangle the mangled name of the canonical type of \p type and
-  /// drill into the Global(TypeMangling(Type())).
-  ///
-  /// \return the child of Type or a nullptr.
-  swift::Demangle::NodePointer
-  DemangleCanonicalType(swift::Demangle::Demangler &Dem,
-                        lldb::opaque_compiler_type_t type);
-
-  /// Return an APINotes manager for the module with module id \id.
-  /// APINotes are used to get at the SDK swiftification annotations.
-  clang::api_notes::APINotesManager *
-  GetAPINotesManager(ClangExternalASTSourceCallbacks *source, unsigned id);
-
-  /// The sibling SwiftASTContext.
-  SwiftASTContext *m_swift_ast_context = nullptr;
-
-  /// The APINotesManager responsible for each Clang module.
-  llvm::DenseMap<clang::Module *,
-                 std::unique_ptr<clang::api_notes::APINotesManager>>
-      m_apinotes_manager;
-};
 
 /// This "middle" class between TypeSystemSwiftTypeRef and
 /// SwiftASTContextForExpressions will eventually go away, as more and
@@ -1200,12 +755,10 @@ public:
   /// Retrieves the modules that need to be implicitly imported in a given
   /// execution scope. This includes the modules imported by both the compile
   /// unit as well as any imports from previous expression evaluations.
-  static bool
-  GetImplicitImports(SwiftASTContext &swift_ast_context, SymbolContext &sc,
-                     ExecutionContextScope &exe_scope,
-                     lldb::StackFrameWP &stack_frame_wp,
-                     llvm::SmallVectorImpl<swift::ModuleDecl *> &modules,
-                     Status &error);
+  static bool GetImplicitImports(
+      SwiftASTContext &swift_ast_context, SymbolContext &sc,
+      ExecutionContextScope &exe_scope, lldb::StackFrameWP &stack_frame_wp,
+      llvm::SmallVectorImpl<swift::ModuleDecl *> &modules, Status &error);
 
   /// Cache the user's imports from a SourceFile in a given execution scope such
   /// that they are carried over into future expression evaluations.
