+++ lldb/source/Plugins/TypeSystem/Swift/TypeSystemSwift.cpp
@@ -1,0 +1,13 @@
+//===-- TypeSystemSwift.cpp ==---------------------------------------------===//
+//
+// This source file is part of the Swift.org open source project
+//
+// Copyright (c) 2014 - 2020 Apple Inc. and the Swift project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See https://swift.org/LICENSE.txt for license information
+// See https://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
+//
+//===----------------------------------------------------------------------===//
+
+#include "Plugins/TypeSystem/Swift/TypeSystemSwift.h"
