+++ lldb/unittests/Symbol/TestSwiftASTContext.cpp
@@ -12,9 +12,9 @@
 
 #include "gtest/gtest.h"
 
+#include "Plugins/TypeSystem/Swift/SwiftASTContext.h"
 #include "lldb/Host/FileSystem.h"
 #include "lldb/Host/HostInfo.h"
-#include "lldb/Symbol/SwiftASTContext.h"
 
 using namespace lldb;
 using namespace lldb_private;
